+++ src/runloop_api_client/resources/devboxes/devboxes.py
@@ -532,7 +532,15 @@ class DevboxesResource(SyncAPIResource):
         limit: int | Omit = omit,
         starting_after: str | Omit = omit,
         status: Literal[
-            "provisioning", "initializing", "running", "suspending", "suspended", "resuming", "failure", "shutdown"
+            "scheduled",
+            "provisioning",
+            "initializing",
+            "running",
+            "suspending",
+            "suspended",
+            "resuming",
+            "failure",
+            "shutdown",
         ]
         | Omit = omit,
         # Use the following arguments if you need to pass additional parameters to the API that aren't available via kwargs.
@@ -600,8 +608,8 @@ class DevboxesResource(SyncAPIResource):
         Devbox.
 
         This tunnel is not persisted on the Devbox and is generated fresh on
-        each request. The returned auth_token must be passed as a Bearer token in the
-        Authorization header.
+        each request. The returned auth_token should be passed as a Bearer token in the
+        X-Runloop-Tunnel-Authorization header.
 
         Args:
           extra_headers: Send extra headers
@@ -2195,7 +2203,15 @@ class AsyncDevboxesResource(AsyncAPIResource):
         limit: int | Omit = omit,
         starting_after: str | Omit = omit,
         status: Literal[
-            "provisioning", "initializing", "running", "suspending", "suspended", "resuming", "failure", "shutdown"
+            "scheduled",
+            "provisioning",
+            "initializing",
+            "running",
+            "suspending",
+            "suspended",
+            "resuming",
+            "failure",
+            "shutdown",
         ]
         | Omit = omit,
         # Use the following arguments if you need to pass additional parameters to the API that aren't available via kwargs.
@@ -2263,8 +2279,8 @@ class AsyncDevboxesResource(AsyncAPIResource):
         Devbox.
 
         This tunnel is not persisted on the Devbox and is generated fresh on
-        each request. The returned auth_token must be passed as a Bearer token in the
-        Authorization header.
+        each request. The returned auth_token should be passed as a Bearer token in the
+        X-Runloop-Tunnel-Authorization header.
 
         Args:
           extra_headers: Send extra headers