+++ tests/api_resources/test_devboxes.py
@@ -250,7 +250,7 @@ class TestDevboxes:
             include_total_count=True,
             limit=0,
             starting_after="starting_after",
-            status="provisioning",
+            status="scheduled",
         )
         assert_matches_type(SyncDevboxesCursorIDPage[DevboxView], devbox, path=["response"])
 
@@ -1919,7 +1919,7 @@ class TestAsyncDevboxes:
             include_total_count=True,
             limit=0,
             starting_after="starting_after",
-            status="provisioning",
+            status="scheduled",
         )
         assert_matches_type(AsyncDevboxesCursorIDPage[DevboxView], devbox, path=["response"])